LayoutTests/compositing/video/video-reflection-expected.txt

EXPECTED (video.currentTime == '1') OK
